chore: rename mask
