+++ src/cli/commands/anon-col/commands/mask/mask-action.ts
