+++ src/cli/commands/anon-col/commands/mask/mask-command.ts
@@ -1,6 +1,7 @@
 import {Command} from 'commander';
 import {defaultMaskConfig} from '../../../../../anonymizers/mask/mask-options';
-import {maskAction} from './action';
+import {addConnectionArgs} from '../../helpers/add-connection-args';
+import {maskAction} from './mask-action';
 
 export const maskCommand = new Command('mask');
 
@@ -12,3 +13,5 @@ maskCommand
 		defaultMaskConfig.character,
 	)
 	.action(maskAction);
+
+addConnectionArgs(maskCommand);
